feat: properly handle cancellation of async notification
To end an async notification, CancelAsyncCall must be called, so that
the refcount on the sink that is kept internally can be released.

There isn't really a better solution than adding a new object that
implements this cancel on drop. Trying to add it to the stream would
have a lifetime bound on it, and would cause issues with the async
queries that are not notifications.